requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
easyocr==1.7.0
opencv-python==4.8.1.78
rapidfuzz==3.5.2
numpy==1.24.3
Pillow==10.0.1
gunicorn==21.2.0